requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
Brotli==1.0.9
-e git+https://github.com/DaMalle/cakebot.git@6825876f1932626e172aea469f0532b3b47bf1c4#egg=cakebot
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.1.0
Deprecated==1.2.13
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
imageio-ffmpeg==0.4.7
multidict==6.0.2
mutagen==1.45.1
numpy==1.23.1
packaging==21.3
py-cord==1.7.3
pycord==0.1.1
pycparser==2.21
pycryptodomex==3.14.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dotenv==0.19.2
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
typing_extensions==4.1.1
urllib3==1.26.10
websockets==10.2
wrapt==1.14.1
yarl==1.7.2
youtube-dl==2021.12.17
yt-dlp==2022.3.8.2
